2023-11-15 0.优化普通资源管理的业务逻辑，修复文件存储bug 1.完善课程资源业务逻辑、封装文件响应函数，包含视频分片和RAW文档下载 2.能够根据教师ID获取其培训计划列表 3.修改教师跟踪进度列表为指定计划下 4.对创建课程部分关闭了自动事务，开启自动事务会导致远程服务外键异常问题

+++ training-resource/src/main/java/com/training/resource/service/Impl/ResourceLessonImpl.java
@@ -1,17 +1,23 @@
 package com.training.resource.service.Impl;
 
 import com.alibaba.fastjson.JSONObject;
+import com.training.common.entity.DataFailRespond;
+import com.training.common.entity.DataRespond;
+import com.training.common.entity.DataSuccessRespond;
 import com.training.common.entity.MsgRespond;
-import com.training.common.entity.result.ChapterInfo;
 import com.training.common.entity.result.LessonInfo;
 import com.training.resource.client.TrainingClient;
 import com.training.resource.entity.request.ResourceLessonReq;
+import com.training.resource.entity.result.ResourceLessonInfo;
 import com.training.resource.entity.table.ResourceLessonTable;
 import com.training.resource.mapper.ResourceLessonMapper;
 import com.training.resource.service.ResourceLessonService;
 import com.training.resource.utils.DataUtil;
+import com.training.resource.utils.FileResUtil;
 import com.training.resource.utils.FileUtil;
 import lombok.AllArgsConstructor;
+import lombok.SneakyThrows;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,6 +28,8 @@ import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 @Service
 @Transactional
@@ -32,19 +40,24 @@ public class ResourceLessonImpl implements ResourceLessonService {
     private final TrainingClient trainingClient;
     private final FileUtil fileUtil;
     private final DataUtil dataUtil;
+    private final FileResUtil fileResUtil;
+    private static final ConcurrentMap<String, String> Path_CACHE = new ConcurrentHashMap<>();
     /**
      * 上传教材资源文件具体实现
      * @param req 教材对象
      * @return 根据处理结果返回消息提示
      * by organwalk 2023-11-03
      */
+    @SneakyThrows
     @Override
     public MsgRespond uploadResourceLesson(ResourceLessonReq req) {
         if (resourceLessonMapper.selectLessonIdByChapterId(req.getChapter_id()) != null){
             return MsgRespond.fail("此课程章节已存在教材资源，无需重复上传");
         }
+
         Integer teacherId = req.getTeacher_id();
         Integer lessonId = req.getLesson_id();
+
         // 检查教师存在性与教师、课程一致性
         String teacherMark = validTeacher(teacherId, lessonId);
         if (Objects.nonNull(teacherMark)){
@@ -55,19 +68,21 @@ public class ResourceLessonImpl implements ResourceLessonService {
         if (Objects.nonNull(chapterMark)){
             return MsgRespond.fail(chapterMark);
         }
-        // 保存教材文件
-        String filePath = fileUtil.getLessonFilePath(teacherId, lessonId, req.getChapter_id(), req.getResource_file());
-        // 保存文件
-        try {
-            File file = new File(filePath);
-            if (!file.getParentFile().exists()) {
-                file.getParentFile().mkdirs(); // 创建路径中所有不存在的目录
-            }
-            req.getResource_file().transferTo(file);
-        } catch (IOException e) {
-            return MsgRespond.fail("内部服务错误，文件上传失败，请稍后再试");
+
+
+        // 检查是否具有相同哈希值的文件
+        String filePath = fileUtil.checkEqualHashFilePath("lesson", req.getFile_hash());
+        if (Objects.isNull(filePath)){
+            filePath = fileUtil.getLessonFilePath(req.getTeacher_id(), req.getLesson_id(), req.getChapter_id(), req.getFile_origin_name());
+            String msg =  uploadVideoLessonResource(filePath, req, "upload");
+            return Objects.equals(msg, "当前文件片段上传成功") || Objects.equals(msg, "教材上传成功")
+                    ? MsgRespond.success(msg)
+                    : MsgRespond.fail(msg);
         }
-        resourceLessonMapper.insertLessonResource(new ResourceLessonTable(null, lessonId, teacherId, req.getChapter_id(), filePath, fileUtil.getFileSaveDateTime()));
+
+        resourceLessonMapper.insertLessonResource(new ResourceLessonTable(null,
+                lessonId, teacherId, req.getChapter_id(), filePath,  fileUtil.getFileSaveDateTime(), req.getFile_hash()));
+
         return MsgRespond.success("教材资源上传成功");
     }
 
@@ -84,23 +99,25 @@ public class ResourceLessonImpl implements ResourceLessonService {
         if (Objects.isNull(idMark)){
             return MsgRespond.fail("重新检查指定的讲师、课程和章节是否正确");
         }
-        // 保存教材文件
-        String filePath = fileUtil.getLessonFilePath(req.getTeacher_id(), req.getLesson_id(), req.getChapter_id(), req.getResource_file());
+
+        String filePath = fileUtil.getLessonFilePath(req.getTeacher_id(), req.getLesson_id(), req.getChapter_id(), req.getFile_origin_name());
         // 获取原文件路径
         String oldFilePath = resourceLessonMapper.selectOldFilePath(req.getLesson_id(), req.getTeacher_id(), req.getChapter_id());
-        // 保存文件
-        try {
-            File file = new File(filePath);
-            req.getResource_file().transferTo(file);
-            // 删除原文件
+        // 检查是否具有两条相同路径的文件
+        Integer pathMark = resourceLessonMapper.selectPathIsOverTwo(oldFilePath);
+        if (pathMark != 2){
+            // 如果不存在，则说明同一份文件不存在多重引用，可删除服务器文件
             File oldFile = new File(oldFilePath);
-            if (oldFile.delete()) {
-                resourceLessonMapper.updateLessonResourcePath(filePath, fileUtil.getFileSaveDateTime(), req.getLesson_id(), req.getTeacher_id(), req.getChapter_id());
+            if (oldFile.exists()){
+                if (oldFile.delete()){
+                    return MsgRespond.fail("文件系统错误，请稍后再试");
+                }
             }
-        } catch (IOException e) {
-            return MsgRespond.fail("内部服务错误，文件上传失败，请稍后再试");
         }
-        return MsgRespond.success("教材资源重传成功");
+        String msg =  uploadVideoLessonResource(filePath, req, "reUpload");
+        return Objects.equals(msg, "当前文件片段上传成功") || Objects.equals(msg, "教材上传成功")
+                ? MsgRespond.success(msg)
+                : MsgRespond.fail(msg);
     }
 
     /**
@@ -113,17 +130,30 @@ public class ResourceLessonImpl implements ResourceLessonService {
      */
     @Override
     public MsgRespond deleteOneLessonResource(Integer teacherId, Integer lessonId, Integer chapterId) {
+
         // 检查课程、章节、教程存在性
         Integer idMark = resourceLessonMapper.selectIdByReUpdateArgs(lessonId, teacherId, chapterId);
         if (Objects.isNull(idMark)){
             return MsgRespond.fail("重新检查指定的讲师、课程和章节是否正确");
         }
-        // 检查教材资源是否存储于服务器中,并删除
-        String filePath = resourceLessonMapper.selectOldFilePath(lessonId, teacherId, chapterId);
-        File file = new File(filePath);
-        if (!file.exists() || file.delete()){
-            resourceLessonMapper.deleteOneLessonResource(teacherId, lessonId, chapterId);
+
+        // 删除数据库记录
+        resourceLessonMapper.deleteOneLessonResource(teacherId, lessonId, chapterId);
+
+        // 检查是否具有相同哈希值的文件
+        String fileHash = resourceLessonMapper.selectFileHashByInfo(teacherId, lessonId, chapterId);
+        String filePath = fileUtil.checkEqualHashFilePath("lesson", fileHash);
+        if (Objects.isNull(filePath)){
+            // 如果不存在相同哈希值的文件，可以直接删除服务器存储记录
+            String oldFilePath = resourceLessonMapper.selectOldFilePath(lessonId, teacherId, chapterId);
+            File file = new File(oldFilePath);
+            if (file.exists()){
+                if (!file.delete()){
+                    return MsgRespond.fail("资源服务器处理错误，请稍后再试");
+                }
+            }
         }
+
         return MsgRespond.success("已成功删除此课程文件");
     }
 
@@ -138,26 +168,73 @@ public class ResourceLessonImpl implements ResourceLessonService {
     public MsgRespond deleteAllLessonResource(Integer teacherId, Integer lessonId) {
         // 检查教师、课程存在性
         Integer idMark = resourceLessonMapper.selectIdByDeleteAllLessonArgs(lessonId, teacherId);
-        if (Objects.isNull(idMark)){
+        if (Objects.isNull(idMark)) {
             return MsgRespond.fail("重新检查指定的讲师、课程是否正确");
         }
-        // 检查教材文件夹是否存储于服务器中，并删除
+
+        // 获取需要删除的路径和不能删除的路径列表
         String folderPath = fileUtil.getLessonFolderPath(teacherId, lessonId);
+        List<String> nonDeletePaths = resourceLessonMapper.findDuplicateFilePathsByTeacherId(teacherId, lessonId);
+
+        // 删除数据库记录
+        resourceLessonMapper.deleteAllLessonResource(teacherId, lessonId);
+
+        // 删除文件
         File file = new File(folderPath);
-        if (!file.exists()){
-            resourceLessonMapper.deleteAllLessonResource(teacherId, lessonId);
-        }else {
+        if (file.exists()) {
             try {
                 Files.walk(Path.of(folderPath))
                         .sorted(Comparator.reverseOrder())
+                        .filter(path -> nonDeletePaths.isEmpty() || !nonDeletePaths.contains(path.toString()))
                         .map(Path::toFile)
                         .forEach(File::delete);
-                resourceLessonMapper.deleteAllLessonResource(teacherId, lessonId);
             } catch (IOException e) {
                 return MsgRespond.fail("内部服务错误，删除失败，请稍后重试");
             }
         }
+
         return MsgRespond.success("已成功删除此课程教材文件");
+    }
+
+    /**
+     * 根据教材资源ID获取教材具体实现
+     * @param rangeString 请求头的range
+     * @param rlId 教材资源ID
+     * @return 根据请求，将返回处理消息、文档教材或视频教材响应体
+     */
+    @SneakyThrows
+    @Override
+    public ResponseEntity<?> getResourceLessonById(String rangeString, Integer rlId) {
+
+        // 检查教材是否存在
+        String filePath = resourceLessonMapper.selectLessonPathById(rlId);
+        if (Objects.isNull(filePath)){
+            return fileResUtil.returnMsg("fail", "未找到教材资源");
+        }
+
+        // 判断属于文档教材还是视频教材
+        String fileExtension = Objects.requireNonNull(filePath).substring(filePath.lastIndexOf("."));
+        if (Objects.equals(fileExtension, ".md")){
+            // 属于文档教材
+            return fileResUtil.returnMarkdown(filePath);
+        }
+        // 属于视频教材
+        return fileResUtil.returnVideo(rangeString, filePath);
+
+    }
+
+    /**
+     * 根据课程ID获取该课程的教材泪飙
+     * @param lessonId 课程ID
+     * @return 返回教材列表，若为空，则返回提示消息
+     */
+    @Override
+    public DataRespond getResourceLessonByLessonId(Integer lessonId) {
+        List<ResourceLessonInfo> resourceLessonList = resourceLessonMapper.selectResourceLessonList(lessonId);
+        if (resourceLessonList.isEmpty()){
+            return new DataFailRespond("该课程尚未上传资源");
+        }
+        return new DataSuccessRespond("已成功获取该课程下的资源列表", resourceLessonList);
     }
 
     /**
@@ -183,4 +260,33 @@ public class ResourceLessonImpl implements ResourceLessonService {
         return null;
     }
 
+    /**
+     * 定义上传视频教材资源的方法
+     * @param filePath 文件路径
+     * @param req 上传请求
+     * @return 处理结果字符串
+     */
+    @SneakyThrows
+    private String uploadVideoLessonResource(String filePath, ResourceLessonReq req, String type){
+        // 获取文件保存路径
+        String savePath = Path_CACHE.get(req.getFile_hash());
+        if (Objects.isNull(savePath)){
+            Path_CACHE.put(req.getFile_hash(), filePath);
+        }
+        String processResult = fileUtil.chunkSaveFile(req.getFile_hash(), filePath, req.getFile_chunks_sum(), req.getFile_now_chunk(), req.getFile_size(), req.getResource_file());
+        if (Objects.isNull(processResult)){
+            return "当前视频教材片段上传成功";
+        }else if (Objects.equals(processResult, "true")){
+            if (Objects.equals(type, "upload")){
+                resourceLessonMapper.insertLessonResource(new ResourceLessonTable(null,
+                        req.getLesson_id(), req.getTeacher_id(), req.getChapter_id(), Path_CACHE.get(req.getFile_hash()),  fileUtil.getFileSaveDateTime(), req.getFile_hash()));
+            }else {
+                resourceLessonMapper.updateLessonResourcePath(Path_CACHE.get(req.getFile_hash()), fileUtil.getFileSaveDateTime(), req.getLesson_id(), req.getTeacher_id(), req.getChapter_id());
+            }
+            Path_CACHE.remove(req.getFile_hash());
+            return "视频教材上传成功";
+        }else {
+            return processResult;
+        }
+    }
 }
